Added Auth Northbricks provider
