+++ config.xml
@@ -77,9 +77,8 @@
         <splash height="960" src="resources/ios/splash/Default@2x~iphone.png" width="640" />
         <splash height="480" src="resources/ios/splash/Default~iphone.png" width="320" />
     </platform>
-    <!-- <allow-navigation href="http://172.20.10.2:8100" /> -->
-    <allow-navigation href="http://localhost:8080/*"/>
-    <allow-navigation href="http://localhost:8100/*"/>
+    <allow-navigation href="http://localhost:8080/*" />
+    <allow-navigation href="http://localhost:8100/*" />
     <engine name="ios" spec="^4.5.1" />
     <plugin name="cordova-plugin-inappbrowser" spec="^1.7.1" />
     <plugin name="cordova-plugin-actionsheet" spec="^2.3.3" />
