+++ ionic.config.json
@@ -1,6 +1,6 @@
 {
   "name": "Northbricks",
-  "app_id": "918f579e",
+  "app_id": "02c7108e",
   "type": "ionic-angular",
   "integrations": {
     "cordova": {}
