+++ src/app/app.module.ts
@@ -20,6 +20,7 @@ import { ToastService } from "../providers/utils/toast.service";
 import { AlertService } from "../providers/utils/alert.service";
 import { LoginPageModule } from "../pages/login/login.module";
 import { AuthServiceProvider } from '../providers/auth-service/auth-service';
+import { AuthServiceNorthbricksProvider } from '../providers/auth-service-northbricks/auth-service-northbricks';
 
 @NgModule({
   declarations: [
@@ -62,6 +63,7 @@ import { AuthServiceProvider } from '../providers/auth-service/auth-service';
     AlertService,
     { provide: ErrorHandler, useClass: IonicErrorHandler },
     AuthServiceProvider,
+    AuthServiceNorthbricksProvider,
   ]
 })
 export class AppModule { }
