+++ src/pages/login/login.html
@@ -11,10 +11,10 @@
   <form (submit)="doLogin()">
     <ion-list>
 
-      <ion-item>
+      <!-- <ion-item>
         <ion-label fixed>Email</ion-label>
         <ion-input type="email" [(ngModel)]="account.email" name="email"></ion-input>
-      </ion-item>
+      </ion-item> -->
 
       <!--
       Want to use a Username instead of an Email? Here you go:
@@ -24,10 +24,10 @@
       </ion-item>
       -->
 
-      <ion-item>
+      <!-- <ion-item>
         <ion-label fixed>Password</ion-label>
         <ion-input type="password" [(ngModel)]="account.password" name="password"></ion-input>
-      </ion-item>
+      </ion-item> -->
 
       <div padding>
         <button ion-button color="primary" block>Login</button>
