+++ src/pages/tabs/tabs.html
@@ -1,5 +1,6 @@
 <ion-tabs #myTabs>
+  <ion-tab [root]="banksRoot" tabTitle="Banks" tabIcon="information-circle"></ion-tab>
   <ion-tab [root]="homeRoot" tabTitle="Transactions" tabIcon="cash"></ion-tab>
-  <!--<ion-tab [root]="tab2Root" tabTitle="About" tabIcon="information-circle"></ion-tab>-->
+
   <ion-tab [root]="profileRoot" tabTitle="About" tabIcon="information-circle"></ion-tab>
 </ion-tabs>
